adaptive_chat/requirements.txt
requests==2.31.0
python-dotenv==1.0.0
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.5.2
tiktoken==0.5.2
langchain==0.1.4
langchain-openai==0.0.5
python-multipart==0.0.6
supervisor>=4.2.5
jsonschema>=4.19.1
supabase>=2.3.0 